+++ CHANGELOG.md
@@ -2,11 +2,12 @@ All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 
-## x.y.z - YYYY-MM-DD
+## Pre-Release 0.0.0 2020-09-14
 
 ### Added
 
-- Lorem ipsum dolor sit amet
+- Initial configuration management
+- Initial calibration workflow (with many TODO's)
 
 ### Deprecated